requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clipman==3.3.1
colorama==0.4.6
cors==1.0.1
dbus-next==0.2.3
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
frozenlist==1.5.0
future==1.0.0
g4f==0.4.4.1
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
# Helpingai-T2==0.5
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
load-dotenv==0.1.0
loguru==0.7.3
markdown-it-py==2.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
orjson==3.10.15
outcome==1.3.0.post0
pillow
poe-api-wrapper==1.7.0
prompt_toolkit==3.0.50
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# python-tgpt==0.8.1
python-vlc==3.0.21203
PyYAML==6.0.1
quickjs==1.19.4
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.3.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tldextract==5.1.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
bcrypt==4.2.1
pytz
Flask-SQLAlchemy
psycopg2-binary
flask-socketio
resend
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client